Fix: No Internet Situations
