+++ app/src/main/java/com/goldenraven/padawanwallet/ui/Composables.kt
@@ -1,12 +1,31 @@
 package com.goldenraven.padawanwallet.ui
 
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.animateColorAsState
+import androidx.compose.animation.expandVertically
+import androidx.compose.animation.shrinkVertically
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
+import com.goldenraven.padawanwallet.R
+import com.goldenraven.padawanwallet.theme.connection_available
+import com.goldenraven.padawanwallet.theme.connection_unavailable
 import com.goldenraven.padawanwallet.theme.md_theme_dark_surfaceLight
+import com.goldenraven.padawanwallet.theme.sofiaPro
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
+import kotlinx.coroutines.delay
 
 @Composable
 internal fun ShowBars() {
@@ -43,4 +62,52 @@ internal fun DrawerScreenAppBarTitle(title: String) {
         style = MaterialTheme.typography.headlineMedium,
         color = MaterialTheme.colorScheme.primary
     )
+}
+
+
+@Composable
+fun ConnectivityStatus(isConnected: Boolean) {
+    var visibility by remember { mutableStateOf(false) }
+
+    AnimatedVisibility(
+        visible = visibility,
+        enter = expandVertically(),
+        exit = shrinkVertically()
+    ) {
+        ConnectivityStatusBox(isConnected = isConnected)
+    }
+
+    LaunchedEffect(isConnected) {
+        visibility = if (!isConnected) {
+            true
+        } else {
+            delay(2000)
+            false
+        }
+    }
+}
+
+@Composable
+fun ConnectivityStatusBox(isConnected: Boolean) {
+    val backgroundColor by animateColorAsState(if (isConnected) connection_available else connection_unavailable)
+    val message = if (isConnected) "Back Online!" else "No Internet Connection!"
+    val iconResource = if (isConnected) {
+        R.drawable.ic_connectivity_available
+    } else {
+        R.drawable.ic_connectivity_unavailable
+    }
+
+    Box(
+        modifier = Modifier
+            .background(backgroundColor)
+            .fillMaxWidth()
+            .padding(8.dp),
+        contentAlignment = Alignment.Center
+    ) {
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            androidx.compose.material.Icon(painterResource(id = iconResource), "Connectivity Icon", tint = Color.White)
+            Spacer(modifier = Modifier.size(8.dp))
+            androidx.compose.material.Text(message, color = Color.White, fontSize = 15.sp)
+        }
+    }
 }
