--- app/src/main/java/com/goldenraven/padawanwallet/utils/isNetworkAvailable.kt
@@ -1,15 +1,0 @@
-/*
- * Copyright 2020-2022 thunderbiscuit and contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the ./LICENSE file.
- */
-
-package com.goldenraven.padawanwallet.utils
-
-import android.content.Context
-import android.net.ConnectivityManager
-
-fun isNetworkAvailable(context: Context): Boolean {
-    val connectionManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-    val activeNetworkInfo = connectionManager.activeNetworkInfo
-    return activeNetworkInfo != null && activeNetworkInfo.isConnectedOrConnecting
-}
